refactor: Obstacle 패키지 이동
