+++ src/main/java/chess/domain/direction/DiagonalStraightDirection.java
@@ -1,6 +1,6 @@
 package chess.domain.direction;
 
-import chess.domain.Obstacle;
+import chess.domain.position.Obstacle;
 import chess.domain.position.Position;
 
 public class DiagonalStraightDirection implements Direction {
