+++ src/main/java/chess/domain/direction/KnightDirection.java
@@ -1,6 +1,6 @@
 package chess.domain.direction;
 
-import chess.domain.Obstacle;
+import chess.domain.position.Obstacle;
 import chess.domain.position.Position;
 import chess.domain.position.Vector;
 import java.util.Arrays;
