+++ src/main/java/chess/domain/direction/SlidingDirection.java
@@ -1,6 +1,6 @@
 package chess.domain.direction;
 
-import chess.domain.Obstacle;
+import chess.domain.position.Obstacle;
 import chess.domain.position.Position;
 import java.util.stream.Stream;
