+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.Obstacle;
+import chess.domain.position.Obstacle;
 import chess.domain.position.Position;
 import java.util.Objects;
