+++ src/test/java/chess/domain/Fixtures.java
@@ -4,6 +4,7 @@ import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
 import chess.domain.position.File;
+import chess.domain.position.Obstacle;
 import chess.domain.position.Position;
 import chess.domain.position.Rank;
 import java.util.List;
