+++ src/test/java/chess/domain/direction/DownRightDirectionTest.java
@@ -6,7 +6,7 @@ import static chess.domain.Fixtures.G2;
 import static chess.domain.Fixtures.H1;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Obstacle;
+import chess.domain.position.Obstacle;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
