+++ src/test/java/chess/domain/direction/RightDirectionTest.java
@@ -7,7 +7,7 @@ import static chess.domain.Fixtures.G3;
 import static chess.domain.Fixtures.H3;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Obstacle;
+import chess.domain.position.Obstacle;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
