+++ src/test/java/chess/domain/direction/UpDirectionTest.java
@@ -11,7 +11,7 @@ import static chess.domain.Fixtures.H1;
 import static chess.domain.Fixtures.H3;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Obstacle;
+import chess.domain.position.Obstacle;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
